BEWARE WITH GLOBALS: eval does not see user's $_ instead it see debegger's one
As you can see from this text output:
  Failed test 'Run debugger commands for all profiles'
  at KP-05-options.t line 60.
--- Got
+++ Expected
@@ -1,8 +1,5 @@
-'"e \\$_"
->>1<<
-"e \\$_"
->>2<<
-"e \\$_"
->>3<<
+'1
+2
+3
 { a => 10 }
  '

The value returned from DB::eval for $_ is 'e $_' instead of expected 1
Also you can see '>>1<<', '>>2<<', '>>3<<' meaning that our change to $_
at DebugHooks:938 does not effect user's one.

If you apply this patch:
--- a/lib/Devel/DebugHooks.pm
+++ b/lib/Devel/DebugHooks.pm
@@ -935,9 +935,10 @@ sub DB {
         # FIX: rename: action -> actions
         if( exists $trap->{ action } ) {
             # Run all actions
-            for( @{ $trap->{ action } } ) {
+            $_ =  'Try to change user\'s $_';
+            for my $action ( @{ $trap->{ action } } ) {
                 # NOTICE: if we do not use scall the $DB::file:$DB::line is broken
-                $ext_call++; scall( \&process, $_ );
+                $ext_call++; scall( \&process, $action );
             }

             # Actions do not stop execution

The test output will be next:

  Failed test 'Run debugger commands for all profiles'
  at ./KP-05-options.t line 60.
--- Got
+++ Expected
@@ -1,8 +1,5 @@
-'"Try to change user\'s \\$_"
->>Try to change user\'s $_<<
-"Try to change user\'s \\$_"
->>Try to change user\'s $_<<
-"Try to change user\'s \\$_"
->>Try to change user\'s $_<<
+'1
+2
+3
 { a => 10 }
 '

As you can see the user's $_ in this case is changed.

So our change to $_ at DebugHooks:938 loop does not affect user's one
But hurt eval executed within a subroutine defined in the DB package which
should see first non-DB piece of code that called it [1]
and not $_ at DebugHooks:938

 [1] http://perldoc.perl.org/functions/eval.html