We should evaluate EXPR in context of package at current position while { dd }
Because we are stopped at debugger we want to see the debugger context, not
user's. And when we access &DB::state we want to see the debugger context too.
But &DB::state is special and the `DB::state( 'inDB', undef );` was used to
show debugger state we are debugging not terminal debugger instance